+++ src/main/java/org/chenliang/freepark/scheduler/GetCouponsTaskScheduler.java
@@ -9,6 +9,7 @@ import org.chenliang.freepark.model.entity.Member;
 import org.chenliang.freepark.model.rtmap.ProductsResponse;
 import org.chenliang.freepark.model.rtmap.ProductsResponse.Product;
 import org.chenliang.freepark.repository.MemberRepository;
+import org.chenliang.freepark.service.CouponsService;
 import org.chenliang.freepark.service.RtmapService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -20,68 +21,16 @@ import org.springframework.stereotype.Service;
 @Service
 public class GetCouponsTaskScheduler {
 
-  @Value(value = "${coupon.name}")
-  private String couponName;
-
-  @Value(value = "${coupon.number}")
-  private int couponNum;
-
-  @Autowired
-  private ThreadPoolTaskScheduler taskScheduler;
-
   @Autowired
   private MemberRepository memberRepository;
 
   @Autowired
-  private RtmapService rtmapService;
+  private CouponsService couponsService;
 
   @Scheduled(cron = "${coupon.cron}")
   public void scheduleGetCouponsTask() {
     final List<Member> members = memberRepository.findAll();
-    Instant now = Instant.now();
-    Random random = new Random();
-
-    final int productId = getProductId(members.get(0));
-    for (Member member : members) {
-      int delaySeconds = random.nextInt(3600 * 2);
-      Instant startTime = now.plusSeconds(delaySeconds);
-      log.info("Scheduled member {} to buy coupons at {}", member.getMobile(), startTime.toString());
-      taskScheduler.schedule(() -> {
-        try {
-          rtmapService.buy(member, productId, couponNum);
-          member.setCoupons(couponNum);
-          memberRepository.save(member);
-        } catch (RtmapApiException ignored) {
-        } catch (Exception e) {
-          log.info("Get coupons task for member {} failed with unexpected exception", member.getMobile(), e);
-        }
-      }, startTime);
-    }
+    couponsService.buyCoupons(members);
   }
 
-  private int getProductId(Member member) {
-    int startPage = 1;
-    int productId = searchCouponId(member, startPage);
-    if (productId == -1) {
-      throw new RuntimeException("Cannot found free parking coupon product");
-    }
-
-    return productId;
-  }
-
-  private int searchCouponId(Member member, int page) {
-    ProductsResponse response = rtmapService.getProducts(member, page);
-    List<Product> products = response.getData().getList();
-    for (Product product : products) {
-      if (product.getMainInfo().contains(couponName)) {
-        return product.getId();
-      }
-    }
-
-    if ((page + 1) > response.getData().getPages()) {
-      return -1;
-    } else {
-      return searchCouponId(member, page + 1);
-    }
-  }
 }
